Fix: fix declarations of some BLAS and LAPACK functions. (#6637)
* Standardize the declarations of BLAS functions in blas_connector_vector

* standardize the declarations of LAPACK functions

* add some comments

* fix some error

* Restore the mistakenly removed const

+++ source/source_estate/module_dm/cal_edm_tddft.cpp
@@ -29,7 +29,7 @@ void cal_edm_tddft(Parallel_Orbitals& pv,
         // mohan add 2024-03-27
         //! be careful, the type of nloc is 'long'
         //! whether the long type is safe, needs more discussion
-        const long nloc = pv.nloc;
+        const int nloc = pv.nloc;
         const int ncol = pv.ncol;
         const int nrow = pv.nrow;
